+++ src/expressions/boolean.js
@@ -3,36 +3,48 @@ const LimeExpression = require('../structs/expression');
 
 /* ------------------------ division ------------------------ */
 
-// Variable class
-class LimeExpressionVariable extends LimeExpression {
+// Boolean class
+class LimeExpressionBoolean extends LimeExpression {
   // Constructor
-  constructor(lime, key) {
+  constructor(lime, value) {
     // Super from expression class
     super(lime, { isSimple: true });
 
     // Build expression
-    this.key = key;
+    this.raw = value;
   }
 
   /* ------------------------ division ------------------------ */
 
   // Get value
   get value() {
-    if (!this.lime.variables.has(this.key)) {
-      throw new Error('error:undefinedVariable');
-    }
-    return this.lime.variables.get(this.key);
+    return !!this.raw;
   }
 
   /* ------------------------ division ------------------------ */
 
+  // Convert to integer
+  toInteger() {
+    return this.value ? this.lime.direct([1]) : this.lime.direct([0]);
+  }
+
+  // Convert to rational
+  toRational() {
+    return this.toInteger().toRational();
+  }
+
   // Simplify function
   simplify() {
-    return this.value.simplify();
+    return this.lime.build('boolean')(this.value);
+  }
+
+  // Print function
+  print() {
+    return this.value ? 'true' : 'false';
   }
 }
 
 /* ------------------------ division ------------------------ */
 
 // Export
-module.exports = LimeExpressionVariable;
+module.exports = LimeExpressionBoolean;