+++ docs/CHANGELOG.md
@@ -1,13 +1,17 @@
 # Changelog
 
 ### v0.1.1
+- Added expressions: `Boolean`, `Variable`
 - Added functions:
   - Arithmetic: `Decimal`, `Exponent`, `Modulo`, `Positive`, `Negative`
+  - Comparison: `Equal`, `Greater`, `GreaterEqual`, `NotEqual`, `Smaller`, `SmallerEqual`
   - Counting: `Factorial`
+- Modified existing functions to support new expression types
+- Improved engine independence of expression and function tests
 ### v0.1.0
 - Set up engine framework, bundling system using Rollup and testing system using Jest
 - Set up prompt and web examples
-- Added expression: `Integer`, `Rational`
+- Added expressions: `Integer`, `Rational`
 - Added functions:
   - Arithmetic: `Add`, `Divide`, `Multiply`, `Subtract`
   - Structural: `CommonBracket`